% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostedBoxplot.R
\name{boostedBoxplot}
\alias{boostedBoxplot}
\title{Boosted Boxplot}
\usage{
boostedBoxplot(y, x, main = "", labx = NULL, laby = NULL, plot.mean = T,
  text.freq = T, las = 1, ylim = c(0, 0), limitVisibleModalities = 30,
  decreasing = NULL, dynamic = F)
}
\arguments{
\item{y}{a numeric vector.}

\item{x}{a factor vector.}

\item{main}{the title of the graphic.}

\item{labx, laby}{\code{character}, labels of the axis.}

\item{plot.mean}{a boolean to indicate whether the mean value of \code{y} must be plotted for each \code{x} category.}

\item{text.freq}{a boolean to indicate whether the \code{x} categories frequencies must be drawn on boxes.}

\item{las}{1 or 2 to indicate whether the x-axis labels must be respectively horizontal or vertical.}

\item{ylim}{a vector of 2 numeric values to indicate the range of y-axis. Example : \code{c(0,1)}.}

\item{limitVisibleModalities}{an integer to indicate the number of categories to retain for the \code{x} variable.
If the number of categories is higher than \code{limitVisibleModalities}, only categories presenting the highest
and the lowest \code{y} median values are plotted.}

\item{decreasing}{a boolean indicating whether categories order on x-axis according to their corresponding median \code{y} value.}

\item{dynamic}{a boolean. If \code{TRUE}, an AmChart is return and the parameters \code{plot.mean} and \code{test.freq} are ignored.}
}
\value{
If \code{dynamicity} is \code{TRUE} it returns an object of class AmChart, else it returns a classical boxplot result list.
}
\description{
Add useful features to R classical boxplot
}
